Setting up types to be exported
