+++ package.json
@@ -1,8 +1,9 @@
 {
-  "name": "wp-page-cache-control",
+  "name": "@alleyinteractive/wp-page-cache-control",
   "version": "1.0.0",
   "license": "GPL-2.0-or-later",
   "main": "wp-page-cache-control.php",
+  "types": "types/index.d.ts",
   "engines": {
     "node": "16",
     "npm": "8"
