+++ services/providers/default/index.ts
@@ -1,0 +1,1 @@
+export { default as DefaultCacheProvider } from './provider'; // eslint-disable-line import/prefer-default-export
