+++ tsconfig.json
@@ -19,7 +19,7 @@
     "target": "es2020",
     "typeRoots": [
       "./node_modules/@types",
-      "./src/types"
+      "./types"
     ]
   },
   "exclude": [
